+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## [Unreleased]
 
+### New Features
+
+- Added support for `\mathtip{math}{tip}` and `\texttip{math}{tip}` commands.
+  These commands are also supported by MathJax.
+- Added the `\error{}` command which displays its content with a red underline.
+- A specific Compute Engine instance can be associated with a mathfield using
+  `mf.computeEngine = ce`. If none is provided, a default Compute Engine 
+  instance is created when necessary. Setting the property to `null` will 
+  prevent the Compute Engine from being used, but the MathJSON format will not
+  be available.
+
 ### Improvements
 
 - The `window.mathlive` global is now `globalThis[Symbol.for("mathlive")]`. This